+++ Tests/Fixtures/Renderer/Atom/MinimumFeedAuthor.php
@@ -16,7 +16,7 @@ use Brotkrueml\FeedGenerator\Contract\CategoryInterface;
 use Brotkrueml\FeedGenerator\Contract\FeedInterface;
 use Brotkrueml\FeedGenerator\Contract\ImageInterface;
 use Brotkrueml\FeedGenerator\Contract\ItemInterface;
-use Brotkrueml\FeedGenerator\Entity\Author;
+use Brotkrueml\FeedGenerator\ValueObject\Author;
 
 final class MinimumFeedAuthor implements FeedInterface
 {