+++ Tests/Fixtures/Renderer/Rss/FeedImageWithEmptyUri.php
@@ -16,7 +16,7 @@ use Brotkrueml\FeedGenerator\Contract\CategoryInterface;
 use Brotkrueml\FeedGenerator\Contract\FeedInterface;
 use Brotkrueml\FeedGenerator\Contract\ImageInterface;
 use Brotkrueml\FeedGenerator\Contract\ItemInterface;
-use Brotkrueml\FeedGenerator\Entity\Image;
+use Brotkrueml\FeedGenerator\ValueObject\Image;
 
 final class FeedImageWithEmptyUri implements FeedInterface
 {