+++ Tests/Fixtures/Renderer/Rss/FullFeed.php
@@ -16,8 +16,8 @@ use Brotkrueml\FeedGenerator\Contract\CategoryInterface;
 use Brotkrueml\FeedGenerator\Contract\FeedInterface;
 use Brotkrueml\FeedGenerator\Contract\ImageInterface;
 use Brotkrueml\FeedGenerator\Contract\ItemInterface;
-use Brotkrueml\FeedGenerator\Entity\Category;
-use Brotkrueml\FeedGenerator\Entity\Image;
+use Brotkrueml\FeedGenerator\ValueObject\Category;
+use Brotkrueml\FeedGenerator\ValueObject\Image;
 
 final class FullFeed implements FeedInterface
 {