+++ Tests/Fixtures/Renderer/Rss/ItemAttachmentWithEmptyLength.php
@@ -16,8 +16,8 @@ use Brotkrueml\FeedGenerator\Contract\CategoryInterface;
 use Brotkrueml\FeedGenerator\Contract\FeedInterface;
 use Brotkrueml\FeedGenerator\Contract\ImageInterface;
 use Brotkrueml\FeedGenerator\Contract\ItemInterface;
-use Brotkrueml\FeedGenerator\Entity\Attachment;
 use Brotkrueml\FeedGenerator\Entity\Item;
+use Brotkrueml\FeedGenerator\ValueObject\Attachment;
 
 final class ItemAttachmentWithEmptyLength implements FeedInterface
 {