+++ Tests/Unit/ValueObject/ImageTest.php
@@ -9,9 +9,9 @@ declare(strict_types=1);
  * LICENSE.txt file that was distributed with this source code.
  */
 
-namespace Brotkrueml\FeedGenerator\Tests\Unit\Entity;
+namespace Brotkrueml\FeedGenerator\Tests\Unit\ValueObject;
 
-use Brotkrueml\FeedGenerator\Entity\Image;
+use Brotkrueml\FeedGenerator\ValueObject\Image;
 use PHPUnit\Framework\TestCase;
 
 final class ImageTest extends TestCase