+++ CHANGELOG.md
@@ -8,9 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Substitute laminas/laminas-feed and jdecool/jsonfeed with custom implementation
-- Rename Enclosure class to Attachment
+- Rename `Enclosure` class to `Attachment`
 - Allow multiple attachments for JSON feeds
-- Integrate FeedCategoryInterface into FeedInterface
+- Integrate `FeedCategoryInterface` into `FeedInterface`
+- Adjust namespace for classes `Attachment`, `Author`, `Category`, `Image`
 
 ## [0.4.0] - 2022-10-31